app/pyproject.toml
[project]
name = "epstein-osint"
version = "0.1.0"
description = "OSINT Document Analysis & RAG Pipeline"
requires-python = ">=3.10,<3.13"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "PyYAML>=6.0.1",
    "redis>=5.0.0",
    "celery>=5.3.0",
    "aiohttp>=3.9.0",
    "aiosqlite>=0.19.0",
    "tenacity>=8.2.0",
    "pymupdf>=1.23.0",
    "pytesseract>=0.3.10",
    "pdf2image>=1.16.0",
    "Pillow>=10.0.0",
    "surya-ocr>=0.4.0",
    "ffmpeg-python>=0.2.0",
    "openai-whisper>=20231117",
    "chromadb>=0.4.0",
    "neo4j>=5.14.0",
    "sqlalchemy>=2.0.0",
    "pydantic-settings>=2.1.0",
    "langgraph>=0.0.20",
    "crewai>=0.11.0",
    "httpx>=0.26.0",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
    "langchain-text-splitters>=0.0.0",
    "sentence-transformers>=2.2.0",
    "onnxruntime>=1.15.0,<1.24.0",
    "langchain>=1.2.10",
]

[tool.setuptools.packages.find]
where = ["."]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.14.0",
    "fakeredis>=2.21.0",
]

[tool.pytest.ini_options]
minversion = "7.0"
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-ra",
    "--strict-markers",
    "--cov=backend",
    "--cov-report=term-missing",
    "--cov-report=json:../data/telemetry/tests/coverage.json",
    "--cov-report=term-missing",
    "--cov-fail-under=85",
    "--junitxml=../data/telemetry/tests/junit_report.xml",
    "--tb=short",
]

[tool.coverage.run]
source = ["backend"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
    "*/.venv/*",
    "*/migrations/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.mypy]
python_version = "3.10"
strict = true
